# Conflicts:
#	features/merchant/review/src/main/java/com/tokopedia/review/common/util/ReviewConstants.kt
#	features/merchant/review/src/main/java/com/tokopedia/review/feature/createreputation/domain/usecase/ProductrevSubmitReviewUseCase.kt

+++ features/transaction/cart/src/main/java/com/tokopedia/cart/view/adapter/cart/CartItemAdapter.java
@@ -108,7 +108,7 @@ public class CartItemAdapter extends RecyclerView.Adapter<RecyclerView.ViewHolde
 
         void onCartItemCheckChanged(int position, int parentPosition, boolean checked);
 
-        void onWishlistCheckChanged(String productId, int cartId, ImageView imageView);
+        void onWishlistCheckChanged(String productId, Long cartId, ImageView imageView);
 
         void onNeedToRefreshSingleShop(int parentPosition);
